+++ examples/electron/esbuild.mjs
@@ -1,0 +1,34 @@
+/**
+ * This file can be edited to adjust the ESBuild build process.
+ * To reset, delete this file and rerun theia build again.
+ */
+import { browserOptions, watch } from './gen-esbuild.browser.mjs';
+import { nodeOptions } from './gen-esbuild.node.mjs';
+import { electronOptions } from './gen-esbuild.electron.mjs';
+import { exposeModulePlugin } from '@theia/bundle-plugin';
+import esbuild from 'esbuild';
+
+browserOptions.plugins.push(exposeModulePlugin());
+
+const browserContext = await esbuild.context(browserOptions);
+const nodeContext = await esbuild.context(nodeOptions);
+const electronContext = await esbuild.context(electronOptions);
+
+if (watch) {
+    await Promise.all([
+        browserContext.watch(),
+        nodeContext.watch(),
+        electronContext.watch(),
+    ]);
+} else {
+    try {
+        await browserContext.rebuild();
+        await browserContext.dispose();
+        await nodeContext.rebuild();
+        await nodeContext.dispose();
+        await electronContext.rebuild();
+        await electronContext.dispose();
+    } catch {
+        process.exit(1);
+    }
+}